Improvements to codegen, reorg
